requirements.txt
python==3.11.9
flask==3.0.3
pillow==10.4.0
APScheduler==3.10.4
easyocr==1.7.1
transformers==4.44.2
textblob==0.18.0.post0